Interface to persisting TaskItem's
Using an interface makes mocking tests easier and in this case
I am also able to use a custom in memory data store as opposed
to a complete database. I am also able to switch in and out
different ways of storing and recalling TaskItem's.

The tests primarilly cover the in memory repository.