Create Task API endpoint
I have created a new type of TaskItem repository that will
allow me to force errors to be returned which measn I don't need
to simulate invalid TaskItems. This is a benefit to using
interfaces in this manor.

The handler is tested on both the happy and sad paths by using
the new MockErrorTaskItemRepository.